+++ docs/CHANGELOG.rst
@@ -18,6 +18,7 @@ Added
 - Start writing the Change Log and include it in the Documentation
 - Add ``docs`` and ``packaging`` Tox testing environments
 - Add ``dev``, a list of extra requirements for development
+- Add ``save_list`` and ``save_file_list`` functions and console commands
 
 Changed
 -------